requirements.txt
discord.py>=2.3.0
Pillow>=10.0.0
aiohttp>=3.8.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
google-genai